requirements.txt
fastmcp>=0.4.0
mcp>=1.2.0
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.8
beautifulsoup4>=4.12
